Use daily date-based index name suffix
Use a daily date-based index name suffix. Paired with an alias, this
allows for nightly full re-indexing without changing the index name.

After midnight:

 - index.py --incremental will exit because the new day's index does
   not exist yet, or because the index has zero records, or because
   there is a full indexing run in progress.

 - create-index.py will create a new index for the day

 - index.py --full && alias.py will index all records, then point
   the alias at the current day's index

If you have an actual index (not alias) in place, you will need to
remove it or do some creative shuffling before implementing this
process.

(Checking in an old change -- we've run this way for a while now.)

Signed-off-by: Jeff Godin <jgodin@tadl.org>